Discord-Forwardbot/requirements.txt
discord.py-self
audioop-lts